+++ lib/omnyshell_client.dart
@@ -24,6 +24,8 @@ export 'src/application/client/file_transfer.dart'
 export 'src/application/client/line_editor.dart'
     show CommandHistory, LineEditor;
 export 'src/application/client/local_command.dart';
+export 'src/application/client/remote_completion.dart'
+    show remoteCompletionCommand;
 export 'src/application/client/remote_session.dart';
 export 'src/application/client/screen_mode_detector.dart';
 export 'src/application/transfer/transfer_engine.dart'